Crate mpl_gumdrop

Program for distributing tokens efficiently via uploading a Merkle root.

Modules

accounts

An Anchor generated module, providing a set of structs mirroring the structs deriving Accounts, where each field is a Pubkey. This is useful for specifying accounts for a client.

instruction

An Anchor generated module containing the program’s set of instructions, where each method handler in the #[program] mod is associated with a struct defining the input arguments to the method. These should be used directly, when one wants to serialize Anchor instruction data, for example, when speciying instructions on a client.

merkle_distributor

The merkle_distributor program.

merkle_proof

These functions deal with verification of Merkle trees (hash trees). Direct port of https://github.com/OpenZeppelin/openzeppelin-contracts/blob/v3.4.0/contracts/cryptography/MerkleProof.sol

program

Structs

CandyMachine

CandyMachineData

Claim

merkle_distributor::claim accounts.

ClaimCandy

merkle_distributor::claim_candy accounts.

ClaimCount

ClaimEdition

merkle_distributor::claim_edition accounts. Wrapper around MintNewEditionFromMasterEditionViaToken

ClaimStatus

ClaimedEvent

Emitted when tokens are claimed.

CloseDistributor

merkle_distributor::close_distributor accounts.

CloseDistributorTokenAccount

[merkle_distributor::close_distributor_token_acconut] accounts.

MerkleDistributor

State for the account which distributes tokens.

NewDistributor

Accounts for merkle_distributor::new_distributor.

Enums

ErrorCode

Statics

ID

The static program ID

Functions

check_id

Confirms that a given pubkey is equivalent to the program ID

entry

The Anchor codegen exposes a programming model where a user defines a set of methods inside of a #[program] module in a way similar to writing RPC request handlers. The macro then generates a bunch of code wrapping these user defined methods into something that can be executed on Solana.

entrypoint⚠

Safety

id

Returns the program ID

Type Definitions

Result

Anchor generated Result to be used as the return type for the program.